+++ iotedgedev/edge.py
@@ -1,10 +1,15 @@
+from iotedgedev.output import Output
+from iotedgedev.envvars import EnvVars
+from iotedgedev.azurecli import AzureCli
+
+
 class Edge:
-    def __init__(self, envvars, output, azure_cli):
+    def __init__(self, envvars: EnvVars, output: Output, azure_cli: AzureCli):
         self.envvars = envvars
         self.output = output
         self.azure_cli = azure_cli
 
-    def deploy(self, manifest_file):
+    def deploy(self, manifest_file: str):
 
         self.output.header("DEPLOYING CONFIGURATION")
 
@@ -12,5 +17,7 @@ class Edge:
         self.envvars.verify_envvar_has_val("DEVICE_CONNECTION_STRING", self.envvars.DEVICE_CONNECTION_INFO)
         self.envvars.verify_envvar_has_val("DEPLOYMENT_CONFIG_FILE", self.envvars.DEPLOYMENT_CONFIG_FILE)
 
-        if self.azure_cli.set_modules(self.envvars.DEVICE_CONNECTION_INFO.device_id, self.envvars.IOTHUB_CONNECTION_INFO, manifest_file):
+        if self.azure_cli.set_modules(config=manifest_file,
+                                      connection_string=self.envvars.IOTHUB_CONNECTION_INFO,
+                                      device_id=self.envvars.DEVICE_CONNECTION_INFO.device_id):
             self.output.footer("DEPLOYMENT COMPLETE")