+++ iotedgedev/solution.py
@@ -31,9 +31,10 @@ class Solution:
         if(runtime_tag == "1.2"):
             edgeagent_schema_version = "1.1"
             edgehub_schema_version = "1.2"
-        
-        self.utility.copy_from_template_dir(".env.tmp", dir_path, dest_file=".env", replacements={"%EDGE_RUNTIME_VERSION%": runtime_tag, "%EDGEAGENT_SCHEMA_VERSION%": edgeagent_schema_version, "%EDGEHUB_SCHEMA_VERSION%": edgehub_schema_version})
-        
+
+        self.utility.copy_from_template_dir(".env.tmp", dir_path, dest_file=".env", replacements={
+                                            "%EDGE_RUNTIME_VERSION%": runtime_tag, "%EDGEAGENT_SCHEMA_VERSION%": edgeagent_schema_version, "%EDGEHUB_SCHEMA_VERSION%": edgehub_schema_version})
+
         if template == "java":
             mod_cmd = "iotedgedev solution add {0} --template {1} --group-id {2}".format(module, template, group_id)
         else: